move `_setHeaderProperties` to `ResponseBase`
nodes `_setHeaderProperties` was used, as it parse more things (especially links!).
Node and browser header properties are now on par.
