+++ lib/client.js
@@ -313,27 +313,6 @@ function Response(req, options) {
     : null;
 }
 
-/**
- * Set header related properties:
- *
- *   - `.type` the content type without params
- *
- * A response of "Content-Type: text/plain; charset=utf-8"
- * will provide you with a `.type` of "text/plain".
- *
- * @param {Object} header
- * @api private
- */
-
-Response.prototype._setHeaderProperties = function(header){
-  // content-type
-  var ct = header['content-type'] || '';
-  this.type = utils.type(ct);
-
-  // params
-  var obj = utils.params(ct);
-  for (var key in obj) this[key] = obj[key];
-};
 ResponseBase(Response.prototype);
 
 /**
