+++ lib/node/response.js
@@ -104,39 +104,6 @@ Response.prototype.toError = function(){
 };
 
 /**
- * Set header related properties:
- *
- *   - `.type` the content type without params
- *
- * A response of "Content-Type: text/plain; charset=utf-8"
- * will provide you with a `.type` of "text/plain".
- *
- * @param {Object} header
- * @api private
- */
-
-Response.prototype._setHeaderProperties = function(header){
-  // TODO: moar!
-  // TODO: make this a util
-
-  // content-type
-  var ct = header['content-type'] || '';
-
-  // params
-  var params = utils.params(ct);
-  for (var key in params) this[key] = params[key];
-
-  this.type = utils.type(ct);
-
-  // links
-  try {
-    if (header.link) this.links = utils.parseLinks(header.link);
-  } catch (err) {
-    // ignore
-  }
-};
-
-/**
  * Set flags such as `.ok` based on `status`.
  *
  * For example a 2xx response will give you a `.ok` of __true__
