+++ lib/response-base.js
@@ -1,5 +1,11 @@
 
 /**
+ * Module dependencies.
+ */
+
+var utils = require('./utils');
+
+/**
  * Expose `ResponseBase`.
  */
 
@@ -41,3 +47,36 @@ function mixin(obj) {
 ResponseBase.prototype.get = function(field){
     return this.header[field.toLowerCase()];
 };
+
+/**
+ * Set header related properties:
+ *
+ *   - `.type` the content type without params
+ *
+ * A response of "Content-Type: text/plain; charset=utf-8"
+ * will provide you with a `.type` of "text/plain".
+ *
+ * @param {Object} header
+ * @api private
+ */
+
+ResponseBase.prototype._setHeaderProperties = function(header){
+    // TODO: moar!
+    // TODO: make this a util
+
+    // content-type
+    var ct = header['content-type'] || '';
+    this.type = utils.type(ct);
+
+    // params
+    var params = utils.params(ct);
+    for (var key in params) this[key] = params[key];
+
+    // links
+    try {
+        if (header.link) this.links = utils.parseLinks(header.link);
+    } catch (err) {
+        // ignore
+    }
+};
+
